Rename `MockHandler` to `MemoryHandler`.
Since we'll be defaulting to the `MemoryHandler` if no handler is
specified rename it from `MockHandler` to make clear it is a proper
handler.
