+++ example/resources/people.js
@@ -4,7 +4,7 @@ jsonApi.define({
   namespace: "json:api",
   resource: "people",
   description: "Used to attribute work to specific people.",
-  handlers: new jsonApi.MockHandler(),
+  handlers: new jsonApi.MemoryHandler(),
   searchParams: { },
   attributes: {
     firstname: jsonApi.Joi.string().alphanum()
