+++ lib/MemoryHandler.js
