+++ lib/jsonApi.js
@@ -11,7 +11,7 @@ var routes = require("./routes");
 var url = require("url");
 
 jsonApi.Joi = ourJoi.Joi;
-jsonApi.MockHandler = require("./MockHandler");
+jsonApi.MemoryHandler = require("./MemoryHandler");
 
 jsonApi.setConfig = function(apiConfig) {
   jsonApi._apiConfig = apiConfig;
